Fixed bin_to_dec64(), works fine now!
Edit: Doesn't work past 32-bits.
To add: dec_to_bin function.

+++ fn.h
@@ -1,3 +1,3 @@
 long power(int base, int exp);
 long bin_to_dec19(long bin);
-int bin_to_dec64(char bin[]);
+long bin_to_dec64(char bin[]);
